# docs/swagger.yaml
basePath: /[v1]/
definitions:
  github_com_semenovem_portal_internal_rest_view.VehicleShort:
    properties:
      id:
        type: integer
      name:
        type: integer
    type: object
  github_com_semenovem_portal_pkg_fail.Response:
    properties:
      additional_fields:
        additionalProperties: true
        type: object
      code:
        type: integer
      message:
        type: string
      validation_errors:
        items:
          $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.ValidationError'
        type: array
    type: object
  github_com_semenovem_portal_pkg_fail.ValidationError:
    properties:
      message:
        type: string
      path:
        type: string
    type: object
...
  internal_abc_people_controller.createUserForm:
    properties:
      avatar_id:
        type: integer
      expired_at:
        type: string
      firstname:
        type: string
      login:
        type: string
      note:
        type: string
      passwd:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        type: string
      surname:
        type: string
    required:
    - firstname
    - surname
    type: object
  internal_abc_people_controller.freeLoginNameResponse:
    properties:
      free:
        type: boolean
      validate_err:
        type: string
    type: object
  internal_abc_people_controller.userPublicProfileView:
    properties:
      avatar:
        type: string
      firstname:
        type: string
      id:
        type: integer
      surname:
        type: string
    type: object
  internal_abc_store_controller.loadView:
    properties:
      payload:
        type: string
    type: object
  internal_abc_store_controller.storeForm:
    properties:
      payload:
        type: string
      storePath:
        type: string
    required:
    - payload
    - storePath
    type: object
  internal_abc_vehicle_controller.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_semenovem_portal_internal_rest_view.VehicleShort'
        type: array
      total:
        type: integer
    type: object
  internal_abc_vehicle_controller.vehicleView:
    type: object
host: localhost:8080
info:
  contact:
    email: semenovem@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: portal API
  version: "1.0"
paths:
...
  /people:
    post:
      description: |
        `expired_at` в формате `2001-03-24T00:00:00Z`
        введенный login нужно проверить, что он допустим `/people/free-login/:login_name`
      parameters:
      - description: данные создаваемого пользователя
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_abc_people_controller.createUserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_people_controller.userPublicProfileView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Создает пользователя
      tags:
      - people
  /people/:user_id:
    delete:
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no-content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Удаляет пользователя
      tags:
      - people
...
  /people/free-login/:login_name:
    get:
      parameters:
      - description: проверяемый логин
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_people_controller.freeLoginNameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Проверяет, свободен ли указанный логин
      tags:
      - people
...
  /store/:store_path:
    delete:
      parameters:
      - description: code store
        in: path
        name: store_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Удаление
      tags:
      - store
    get:
      parameters:
      - description: code store
        in: path
        name: store_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_store_controller.loadView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Чтение произвольных клиентских данных
      tags:
      - store
    post:
      description: |
        Для возможности восстановления состоянии на клиенте
      parameters:
      - description: code store
        in: path
        name: store_code
        required: true
        type: string
      - description: Данные для сохранения
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_abc_store_controller.storeForm'
      produces:
      - application/json
      responses:
        "201":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Сохранение произвольных клиентских данных
      tags:
      - store
  /vehicles:
    get:
      parameters:
      - in: query
        name: end_time
        type: string
      - in: query
        maximum: 100
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: slug[]
        type: array
      - in: query
        name: start_time
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: user_id[]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_vehicle_controller.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Ищет автомобили по фильтру
      tags:
      - vehicles
  /vehicles/:vehicle_id:
    get:
      parameters:
      - description: ID автомобиля
        in: path
        name: vehicle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_vehicle_controller.vehicleView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Получает данные автомобиля по ID
      tags:
      - vehicles
    post:
      parameters:
      - description: ID автомобиля
        in: path
        name: vehicle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Удаляет автомобиль
      tags:
      - vehicles
    put:
      parameters:
      - description: ID автомобиля
        in: path
        name: vehicle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_vehicle_controller.vehicleView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Обновляет данные автомобиля
      tags:
      - vehicles
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"